# Creating a baseline performance profile