--- a/src/gamelogic/index.ts
+++ b/src/gamelogic/index.ts
@@ -174,7 +174,7 @@ export default class MinesweeperGame {
   private _pushGridChange(x: number, y: number) {
     this._gridChanges.push([x, y]);
     if (this._gridChanges.length >= MinesweeperGame.EMIT_THRESHOLD) {
-      this._emit();
+      // this._emit();
     }
   }