@jakearchibald 2019-04-10
I've removed the iterator from https://github.com/GoogleChromeLabs/graviton/compare/ui-work...optimise-game-logic. Going to do more optimisation work there