@PaulKinlan 2019-04-10
Time based chunking breaking at 60ms intervals

![image](https://user-images.githubusercontent.com/45510/55885935-bc178b80-5ba2-11e9-9fd4-b02bbb7c9845.png)